requirements.txt
highrise-bot-sdk
deep-translator
python-dotenv